The main layout of the web site has been changed
I modified the layout `page`, in particular the navigation division.
Now, the navigation division has been placed to the left sticky side
bar. The outfit of navigation links is rought and will have to be
developed in the next future.

The new layout has been added to the `_include` directory into the
`sidebar_left.html` file